requirements.txt
open3d 
POT 
dtw
dynamixel_sdk 
numpy
pandas 
keyboard 
bs4 
isort
scikit-learn
hydra-core
tqdm 
matplotlib
wandb
diffusers==0.11.1
einops
huggingface_hub==0.11.1
opencv-python
imageio[ffmpeg]
zmq
blosc
h5py
flask
shapely
opencv-contrib-python
record3d
